python/MyChatbot/Boto/requirements.txt
transformers
langchain
langchain-community
pinecone-client
langchain-pinecone
langchain_ollama
sentence-transformers
colab-xterm
requests
tqdm

